cleanup: move more classes into correct package
