+++ dubbo-config/dubbo-config-api/src/test/java/com/alibaba/dubbo/config/AbstractInterfaceConfigTest.java
@@ -19,6 +19,12 @@ package com.alibaba.dubbo.config;
 import com.alibaba.dubbo.common.Constants;
 import com.alibaba.dubbo.common.URL;
 import com.alibaba.dubbo.common.utils.ConfigUtils;
+import com.alibaba.dubbo.config.api.Greeting;
+import com.alibaba.dubbo.config.mock.GreetingLocal1;
+import com.alibaba.dubbo.config.mock.GreetingLocal2;
+import com.alibaba.dubbo.config.mock.GreetingLocal3;
+import com.alibaba.dubbo.config.mock.GreetingMock1;
+import com.alibaba.dubbo.config.mock.GreetingMock2;
 import com.alibaba.dubbo.monitor.MonitorService;
 import com.alibaba.dubbo.registry.RegistryService;
 import junit.framework.TestCase;
