+++ dubbo-config/dubbo-config-api/src/test/java/com/alibaba/dubbo/config/mock/GreetingLocal2.java
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,9 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.config;
+package com.alibaba.dubbo.config.mock;
 
-class GreetingLocal2 implements Greeting {
+import com.alibaba.dubbo.config.api.Greeting;
+
+public class GreetingLocal2 implements Greeting {
     @Override
     public String hello() {
         return "local";
