+++ dubbo-config/dubbo-config-api/src/test/java/com/alibaba/dubbo/config/mock/GreetingMock1.java
@@ -6,7 +6,7 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,17 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.config;
+package com.alibaba.dubbo.config.mock;
 
-class GreetingLocal3 implements Greeting {
-    private Greeting greeting;
-
-    public GreetingLocal3(Greeting greeting) {
-        this.greeting = greeting;
-    }
-
-    @Override
-    public String hello() {
-        return null;
-    }
+public class GreetingMock1 {
 }
